@@ -5,35 +5,38 @@
 //  2018 Fabrizio Duroni.
 //
 
-import XCTest
+import Testing
 @testable import ID3TagEditor
 
-class ID3FrameContentSizeCalculatorTest: XCTestCase {
-    private let id3FrameContentSizeCalculator = ID3FrameContentSizeCalculator(
-            uInt32ToByteArrayAdapter: MockUInt32ToByteArrayAdapter(),
-            synchsafeEncoder: SynchsafeIntegerEncoder()
-    )
+struct ID3FrameContentSizeCalculatorTest {
 
-    func testContentCalculateSizeForVersion2() {
+    @Test func testContentCalculateSizeForVersion2() {
+        let id3FrameContentSizeCalculator = ID3FrameContentSizeCalculator(
+                uInt32ToByteArrayAdapter: MockUInt32ToByteArrayAdapter(),
+                synchsafeEncoder: SynchsafeIntegerEncoder()
+        )
         let size = id3FrameContentSizeCalculator.calculateSizeOf(content: [0x1, 0x1], version: .version2)
 
-        XCTAssertEqual(size, [0x2])
+        #expect(size == [0x2])
     }
 
-    func testContentCalculateSizeForVersion3() {
+    @Test func testContentCalculateSizeForVersion3() {
+        let id3FrameContentSizeCalculator = ID3FrameContentSizeCalculator(
+                uInt32ToByteArrayAdapter: MockUInt32ToByteArrayAdapter(),
+                synchsafeEncoder: SynchsafeIntegerEncoder()
+        )
         let size = id3FrameContentSizeCalculator.calculateSizeOf(content: [0x1, 0x1], version: .version3)
 
-        XCTAssertEqual(size, [0x3, 0x2])
+        #expect(size == [0x3, 0x2])
     }
 
-    func testContentCalculateSizeForVersion4() {
+    @Test func testContentCalculateSizeForVersion4() {
+        let id3FrameContentSizeCalculator = ID3FrameContentSizeCalculator(
+                uInt32ToByteArrayAdapter: MockUInt32ToByteArrayAdapter(),
+                synchsafeEncoder: SynchsafeIntegerEncoder()
+        )
         let size = id3FrameContentSizeCalculator.calculateSizeOf(content: [0x1, 0x1], version: .version4)
 
-        XCTAssertEqual(size, [0x3, 0x2])
+        #expect(size == [0x3, 0x2])
     }
-
-    static let allTests = [
-        ("testContentCalculateSizeForVersion2", testContentCalculateSizeForVersion2),
-        ("testContentCalculateSizeForVersion3", testContentCalculateSizeForVersion3)
-    ]
 }
