@@ -6,21 +6,17 @@
 //  2020 Fabrizio Duroni.
 //
 
-import XCTest
+import Testing
 @testable import ID3TagEditor
 
-class ID3FrameHeaderCreatorTest: XCTestCase {
-    static let allTests = [
-        ("testCreate", testCreate)
-    ]
-
-    func testCreate() {
+struct ID3FrameHeaderCreatorTest {
+    @Test func create() {
         let headerCreator = ID3FrameHeaderCreator(id3FrameConfiguration: ID3FrameConfiguration(),
                                                   frameContentSizeCalculator: MockFrameContentSizeCalculatorForHeader(),
                                                   frameFlagsCreator: MockFrameFlagsCreator())
 
         let header = headerCreator.createUsing(version: .version3, frameType: .unsyncronisedLyrics, frameBody: [0x11])
 
-        XCTAssertEqual(header, [UInt8]("USLT".utf8) + [0x11, 0x00])
+        #expect(header == [UInt8]("USLT".utf8) + [0x11, 0x00])
     }
 }
