@@ -6,11 +6,11 @@
 //  Copyright © 2020 Fabrizio Duroni. All rights reserved.
 //
 
-import XCTest
+import Testing
 @testable import ID3TagEditor
 
-class ID3LocalizedFrameCreatorTest: XCTestCase {
-    func testCreatev2() {
+struct ID3LocalizedFrameCreatorTest {
+    @Test func testCreatev2() {
         let creator = ID3LocalizedFrameCreator(
             id3FrameConfiguration: ID3FrameConfiguration(),
             frameHeaderCreator: MockFrameHeaderCreator(),
@@ -23,7 +23,7 @@ class ID3LocalizedFrameCreatorTest: XCTestCase {
                                         version: .version2,
                                         frameType: .unsyncronisedLyrics)
 
-        XCTAssertEqual(frame,
+        #expect(frame ==
                        [0x01, 0x02, 0x01, 0x69, 0x74, 0x61, 0xFF, 0xFE, 0x63, 0x00, 0x6F, 0x00,
                         0x6E, 0x00, 0x74, 0x00, 0x65, 0x00, 0x6E, 0x00, 0x74, 0x00, 0x20, 0x00,
                         0x64, 0x00, 0x65, 0x00, 0x73, 0x00, 0x63, 0x00, 0x72, 0x00, 0x69, 0x00,
@@ -32,7 +32,7 @@ class ID3LocalizedFrameCreatorTest: XCTestCase {
                         0x00])
     }
 
-    func testCreatev3() {
+    @Test func testCreatev3() {
         let creator = ID3LocalizedFrameCreator(
             id3FrameConfiguration: ID3FrameConfiguration(),
             frameHeaderCreator: MockFrameHeaderCreator(),
@@ -45,7 +45,7 @@ class ID3LocalizedFrameCreatorTest: XCTestCase {
                                         version: .version3,
                                         frameType: .unsyncronisedLyrics)
 
-        XCTAssertEqual(frame,
+        #expect(frame ==
                        [0x01, 0x02, 0x01, 0x69, 0x74, 0x61, 0xFF, 0xFE, 0x63, 0x00, 0x6F, 0x00,
                         0x6E, 0x00, 0x74, 0x00, 0x65, 0x00, 0x6E, 0x00, 0x74, 0x00, 0x20, 0x00,
                         0x64, 0x00, 0x65, 0x00, 0x73, 0x00, 0x63, 0x00, 0x72, 0x00, 0x69, 0x00,
@@ -54,7 +54,7 @@ class ID3LocalizedFrameCreatorTest: XCTestCase {
                         0x00])
     }
 
-    func testCreatev4() {
+    @Test func testCreatev4() {
         let creator = ID3LocalizedFrameCreator(
             id3FrameConfiguration: ID3FrameConfiguration(),
             frameHeaderCreator: MockFrameHeaderCreator(),
@@ -67,7 +67,7 @@ class ID3LocalizedFrameCreatorTest: XCTestCase {
                                         version: .version4,
                                         frameType: .unsyncronisedLyrics)
 
-        XCTAssertEqual(frame,
+        #expect(frame ==
                        [0x01, 0x02, 0x01, 0x69, 0x74, 0x61, 0xFF, 0xFE, 0x63, 0x00, 0x6F, 0x00,
                         0x6E, 0x00, 0x74, 0x00, 0x65, 0x00, 0x6E, 0x00, 0x74, 0x00, 0x20, 0x00,
                         0x64, 0x00, 0x65, 0x00, 0x73, 0x00, 0x63, 0x00, 0x72, 0x00, 0x69, 0x00,
@@ -75,10 +75,4 @@ class ID3LocalizedFrameCreatorTest: XCTestCase {
                         0xFE, 0x6C, 0x00, 0x79, 0x00, 0x72, 0x00, 0x69, 0x00, 0x63, 0x00, 0x73,
                         0x00])
     }
-
-    static let allTests = [
-        ("testCreatev2", testCreatev2),
-        ("testCreatev3", testCreatev3),
-        ("testCreatev4", testCreatev4)
-    ]
 }
