@@ -5,15 +5,11 @@
 //  2018 Fabrizio Duroni.
 //
 
-import XCTest
+import Testing
 @testable import ID3TagEditor
 
-class ID3FrameConfigurationTest: XCTestCase {
-    func testInvalidIdentifierName() {
-        XCTAssertEqual(ID3FrameConfiguration().frameTypeFor(identifier: "::not-valid::", version: .version2), .invalid)
+struct ID3FrameConfigurationTest {
+    @Test func invalidIdentifierName() {
+        #expect(ID3FrameConfiguration().frameTypeFor(identifier: "::not-valid::", version: .version2) == .invalid)
     }
-
-    static let allTests = [
-        ("testInvalidIdentifierName", testInvalidIdentifierName)
-    ]
 }
