@@ -5,47 +5,50 @@
 //  2018 Fabrizio Duroni.
 //
 
-import XCTest
+import Foundation
+import Testing
+
 @testable import ID3TagEditor
 
-class ID3RecordingTimeFrameContentParsingOperationTest: XCTestCase {
-    func testframeContentValidTimestamp() {
-        let expectation = XCTestExpectation(description: "recording time")
+struct ID3RecordingTimeFrameContentParsingOperationTest {
+    @Test func testframeContentValidTimestamp() async {
         let recordingTimeParsingOperation = ID3RecordingTimeFrameContentParsingOperation(
             stringContentParser: ID3FrameStringContentParserFactory.make()
         )
 
-        recordingTimeParsingOperation.parse(frame: frameV4utf8Valid(),
-                                            version: .version4,
-                                            completed: {(frameName, frame) in
-            XCTAssertEqual(frameName, .recordingDateTime)
-            XCTAssertEqual((frame as? ID3FrameRecordingDateTime)?.recordingDateTime.date?.day, 27)
-            XCTAssertEqual((frame as? ID3FrameRecordingDateTime)?.recordingDateTime.date?.month, 7)
-            XCTAssertEqual((frame as? ID3FrameRecordingDateTime)?.recordingDateTime.date?.year, 2018)
-            XCTAssertEqual((frame as? ID3FrameRecordingDateTime)?.recordingDateTime.time?.hour, 11)
-            XCTAssertEqual((frame as? ID3FrameRecordingDateTime)?.recordingDateTime.time?.minute, 35)
-            expectation.fulfill()
-
-        })
+        await confirmation("recording time") { recordingTime in
+            recordingTimeParsingOperation.parse(frame: frameV4utf8Valid(),
+                                                version: .version4,
+                                                completed: {(frameName, frame) in
+                #expect(frameName == .recordingDateTime)
+                #expect((frame as? ID3FrameRecordingDateTime)?.recordingDateTime.date?.day == 27)
+                #expect((frame as? ID3FrameRecordingDateTime)?.recordingDateTime.date?.month == 7)
+                #expect((frame as? ID3FrameRecordingDateTime)?.recordingDateTime.date?.year == 2018)
+                #expect((frame as? ID3FrameRecordingDateTime)?.recordingDateTime.time?.hour == 11)
+                #expect((frame as? ID3FrameRecordingDateTime)?.recordingDateTime.time?.minute == 35)
+                recordingTime()
+            })
+        }
     }
 
-    func testframeContentWithYearInsteadOfTimestamp() {
-        let expectation = XCTestExpectation(description: "recording time")
+    @Test func testframeContentWithYearInsteadOfTimestamp() async {
         let recordingTimeParsingOperation = ID3RecordingTimeFrameContentParsingOperation(
             stringContentParser: ID3FrameStringContentParserFactory.make()
         )
 
-        recordingTimeParsingOperation.parse(frame: frameV4utf8Invalid(),
-                                            version: .version4,
-                                            completed: {(frameName, frame) in
-            XCTAssertEqual(frameName, .recordingDateTime)
-            XCTAssertNil((frame as? ID3FrameRecordingDateTime)?.recordingDateTime.date?.day)
-            XCTAssertNil((frame as? ID3FrameRecordingDateTime)?.recordingDateTime.date?.month)
-            XCTAssertEqual((frame as? ID3FrameRecordingDateTime)?.recordingDateTime.date?.year, 2016)
-            XCTAssertNil((frame as? ID3FrameRecordingDateTime)?.recordingDateTime.time?.hour)
-            XCTAssertNil((frame as? ID3FrameRecordingDateTime)?.recordingDateTime.time?.minute)
-            expectation.fulfill()
-        })
+        await confirmation("recording time") { recordingTime in
+            recordingTimeParsingOperation.parse(frame: frameV4utf8Invalid(),
+                                                version: .version4,
+                                                completed: {(frameName, frame) in
+                #expect(frameName == .recordingDateTime)
+                #expect((frame as? ID3FrameRecordingDateTime)?.recordingDateTime.date?.day == nil)
+                #expect((frame as? ID3FrameRecordingDateTime)?.recordingDateTime.date?.month == nil)
+                #expect((frame as? ID3FrameRecordingDateTime)?.recordingDateTime.date?.year == 2016)
+                #expect((frame as? ID3FrameRecordingDateTime)?.recordingDateTime.time?.hour == nil)
+                #expect((frame as? ID3FrameRecordingDateTime)?.recordingDateTime.time?.minute == nil)
+                recordingTime()
+            })
+        }
     }
 
     private func frameV4utf8Valid() -> Data {
@@ -61,9 +64,4 @@ class ID3RecordingTimeFrameContentParsingOperationTest: XCTestCase {
                 [0x03] + [UInt8]("2016".utf8)
         )
     }
-
-    static let allTests = [
-        ("testframeContentValidTimestamp", testframeContentValidTimestamp),
-        ("testframeContentWithYearInsteadOfTimestamp", testframeContentWithYearInsteadOfTimestamp)
-    ]
 }
