@@ -5,104 +5,156 @@
 //  2018 Fabrizio Duroni.
 //
 
-import XCTest
+import Foundation
+import Testing
+
 @testable import ID3TagEditor
 
-class ID3StringContentParsingOperationTest: XCTestCase {
-    private let stringContentParser = ID3FrameStringContentParser(
-        stringEncodingDetector: ID3FrameStringEncodingDetector(
-            id3FrameConfiguration: ID3FrameConfiguration(),
-            id3StringEncodingConverter: ID3StringEncodingConverter()
-        ),
-        paddingRemover: PaddingRemoverUsingTrimming(),
-        id3FrameConfiguration: ID3FrameConfiguration()
-    )
-
-    func testTooSmallDataValue() {
+struct ID3StringContentParsingOperationTest {
+    @Test func testTooSmallDataValue() {
+        let stringContentParser = ID3FrameStringContentParser(
+            stringEncodingDetector: ID3FrameStringEncodingDetector(
+                id3FrameConfiguration: ID3FrameConfiguration(),
+                id3StringEncodingConverter: ID3StringEncodingConverter()
+            ),
+            paddingRemover: PaddingRemoverUsingTrimming(),
+            id3FrameConfiguration: ID3FrameConfiguration()
+        )
         let value = stringContentParser.parse(frame: Data(capacity: 1), version: .version3)
-        XCTAssertNil(value)
+        #expect(value == nil)
     }
 
-    func testFrameContentParsedV2() {
-        let expectation = XCTestExpectation(description: "content without padding")
+    @Test func testFrameContentParsedV2() async {
+        let stringContentParser = ID3FrameStringContentParser(
+            stringEncodingDetector: ID3FrameStringEncodingDetector(
+                id3FrameConfiguration: ID3FrameConfiguration(),
+                id3StringEncodingConverter: ID3StringEncodingConverter()
+            ),
+            paddingRemover: PaddingRemoverUsingTrimming(),
+            id3FrameConfiguration: ID3FrameConfiguration()
+        )
         let id3StringContentParsingOperation = ID3FrameStringContentParsingOperation(
             stringContentParser: stringContentParser
         ) { (_: String) in
             return (.title, ID3FrameWithStringContent(content: ":: value ::"))
         }
 
-        id3StringContentParsingOperation.parse(frame: frameV2(), version: .version2, completed: {(frameName, frame) in
-            XCTAssertEqual(frameName, .title)
-            XCTAssertEqual((frame as? ID3FrameWithStringContent)?.content, ":: value ::")
-            expectation.fulfill()
-        })
-    }
-
-    func testFrameContentParsedV3() {
-        let expectation = XCTestExpectation(description: "content without padding")
-        let id3StringContentParsingOperation = ID3FrameStringContentParsingOperation(
-            stringContentParser: stringContentParser
-        ) { (_: String) in
-            return (.title, ID3FrameWithStringContent(content: ":: value ::"))
-        }
-
-        id3StringContentParsingOperation.parse(frame: frameV3(), version: .version2, completed: {(frameName, frame) in
-            XCTAssertEqual(frameName, .title)
-            XCTAssertEqual((frame as? ID3FrameWithStringContent)?.content, ":: value ::")
-            expectation.fulfill()
-        })
-    }
-
-    func testFrameContentParsedV2utf16() {
-        let expectation = XCTestExpectation(description: "content without padding")
-        let id3StringContentParsingOperation = ID3FrameStringContentParsingOperation(
-            stringContentParser: stringContentParser
-        ) { (_: String) in
-            return (.title, ID3FrameWithStringContent(content: ":: π value ::"))
-        }
-
-        id3StringContentParsingOperation.parse(
-            frame: frameV2utf16(), version: .version2,
-            completed: {(frameName, frame) in
-                XCTAssertEqual(frameName, .title)
-                XCTAssertEqual((frame as? ID3FrameWithStringContent)?.content, ":: π value ::")
-                expectation.fulfill()
+        await confirmation("content without padding") { fulfill in
+            id3StringContentParsingOperation.parse(
+                frame: frameV2(),
+                version: .version2,
+                completed: {(frameName, frame) in
+                    #expect(frameName == .title)
+                    #expect((frame as? ID3FrameWithStringContent)?.content == ":: value ::")
+                    fulfill()
             })
+        }
     }
 
-    func testFrameContentParsedV3utf16() {
-        let expectation = XCTestExpectation(description: "content without padding")
+    @Test func testFrameContentParsedV3() async {
+        let stringContentParser = ID3FrameStringContentParser(
+            stringEncodingDetector: ID3FrameStringEncodingDetector(
+                id3FrameConfiguration: ID3FrameConfiguration(),
+                id3StringEncodingConverter: ID3StringEncodingConverter()
+            ),
+            paddingRemover: PaddingRemoverUsingTrimming(),
+            id3FrameConfiguration: ID3FrameConfiguration()
+        )
+        let id3StringContentParsingOperation = ID3FrameStringContentParsingOperation(
+            stringContentParser: stringContentParser
+        ) { (_: String) in
+            return (.title, ID3FrameWithStringContent(content: ":: value ::"))
+        }
+
+        await confirmation("content without padding") { fulfill in
+            id3StringContentParsingOperation.parse(
+                frame: frameV3(),
+                version: .version2,
+                completed: {(frameName, frame) in
+                    #expect(frameName == .title)
+                    #expect((frame as? ID3FrameWithStringContent)?.content == ":: value ::")
+                    fulfill()
+            })
+        }
+    }
+
+    @Test func testFrameContentParsedV2utf16() async {
+        let stringContentParser = ID3FrameStringContentParser(
+            stringEncodingDetector: ID3FrameStringEncodingDetector(
+                id3FrameConfiguration: ID3FrameConfiguration(),
+                id3StringEncodingConverter: ID3StringEncodingConverter()
+            ),
+            paddingRemover: PaddingRemoverUsingTrimming(),
+            id3FrameConfiguration: ID3FrameConfiguration()
+        )
         let id3StringContentParsingOperation = ID3FrameStringContentParsingOperation(
             stringContentParser: stringContentParser
         ) { (_: String) in
             return (.title, ID3FrameWithStringContent(content: ":: π value ::"))
         }
 
-        id3StringContentParsingOperation.parse(
-            frame: frameV3utf16(), version: .version3,
-            completed: {(frameName, frame) in
-                XCTAssertEqual(frameName, .title)
-                XCTAssertEqual((frame as? ID3FrameWithStringContent)?.content, ":: π value ::")
-                expectation.fulfill()
-        })
+        await confirmation("content without padding") { fulfill in
+            id3StringContentParsingOperation.parse(
+                frame: frameV2utf16(), version: .version2,
+                completed: {(frameName, frame) in
+                    #expect(frameName == .title)
+                    #expect((frame as? ID3FrameWithStringContent)?.content == ":: π value ::")
+                    fulfill()
+                })
+        }
     }
 
-    func testFrameContentParsedV4utf8() {
-        let expectation = XCTestExpectation(description: "content without padding")
+    @Test func testFrameContentParsedV3utf16() async {
+        let stringContentParser = ID3FrameStringContentParser(
+            stringEncodingDetector: ID3FrameStringEncodingDetector(
+                id3FrameConfiguration: ID3FrameConfiguration(),
+                id3StringEncodingConverter: ID3StringEncodingConverter()
+            ),
+            paddingRemover: PaddingRemoverUsingTrimming(),
+            id3FrameConfiguration: ID3FrameConfiguration()
+        )
         let id3StringContentParsingOperation = ID3FrameStringContentParsingOperation(
             stringContentParser: stringContentParser
         ) { (_: String) in
             return (.title, ID3FrameWithStringContent(content: ":: π value ::"))
         }
 
-        id3StringContentParsingOperation.parse(
-            frame: frameV4utf8(),
-            version: .version4,
-            completed: {(frameName, frame) in
-                XCTAssertEqual(frameName, .title)
-                XCTAssertEqual((frame as? ID3FrameWithStringContent)?.content, ":: π value ::")
-                expectation.fulfill()
-        })
+        await confirmation("content without padding") { fulfill in
+            id3StringContentParsingOperation.parse(
+                frame: frameV3utf16(), version: .version3,
+                completed: {(frameName, frame) in
+                    #expect(frameName == .title)
+                    #expect((frame as? ID3FrameWithStringContent)?.content == ":: π value ::")
+                    fulfill()
+                })
+        }
+    }
+
+    @Test func testFrameContentParsedV4utf8() async {
+        let stringContentParser = ID3FrameStringContentParser(
+            stringEncodingDetector: ID3FrameStringEncodingDetector(
+                id3FrameConfiguration: ID3FrameConfiguration(),
+                id3StringEncodingConverter: ID3StringEncodingConverter()
+            ),
+            paddingRemover: PaddingRemoverUsingTrimming(),
+            id3FrameConfiguration: ID3FrameConfiguration()
+        )
+        let id3StringContentParsingOperation = ID3FrameStringContentParsingOperation(
+            stringContentParser: stringContentParser
+        ) { (_: String) in
+            return (.title, ID3FrameWithStringContent(content: ":: π value ::"))
+        }
+
+        await confirmation("content without padding") { fulfill in
+            id3StringContentParsingOperation.parse(
+                frame: frameV4utf8(),
+                version: .version4,
+                completed: {(frameName, frame) in
+                    #expect(frameName == .title)
+                    #expect((frame as? ID3FrameWithStringContent)?.content == ":: π value ::")
+                    fulfill()
+                })
+        }
     }
 
     private func frameV2() -> Data {
@@ -138,12 +190,4 @@ class ID3StringContentParsingOperationTest: XCTestCase {
                 [0x03] + [UInt8](":: π value ::".utf8)
         )
     }
-
-    static let allTests = [
-        ("testFrameContentParsedV2", testFrameContentParsedV2),
-        ("testFrameContentParsedV3", testFrameContentParsedV3),
-        ("testFrameContentParsedV2utf16", testFrameContentParsedV2utf16),
-        ("testFrameContentParsedV3utf16", testFrameContentParsedV3utf16),
-        ("testFrameContentParsedV4utf8", testFrameContentParsedV4utf8)
-    ]
 }
