@@ -8,20 +8,25 @@
 
 import Foundation
 
-func enumerateLocalizedFrameName(frameName: (ID3FrameContentLanguage) -> FrameName) -> [FrameName] {
-    return ID3FrameContentLanguage.allCases.map({ frameName($0) })
-}
+struct FrameNamesWithLocalizedContent {
+    func get() -> [FrameType: [FrameName]] {
+        return [
+            FrameType.unsyncronisedLyrics: enumerateLocalizedFrameName(frameName: FrameName.unsynchronizedLyrics),
+            FrameType.comment: enumerateLocalizedFrameName(frameName: FrameName.comment)
+        ]
+    }
 
-let frameNamesWithLocalizedContent = [
-    FrameType.unsyncronisedLyrics: enumerateLocalizedFrameName(frameName: FrameName.unsynchronizedLyrics),
-    FrameType.comment: enumerateLocalizedFrameName(frameName: FrameName.comment)
-]
+    private func enumerateLocalizedFrameName(frameName: (ID3FrameContentLanguage) -> FrameName) -> [FrameName] {
+        return ID3FrameContentLanguage.allCases.map({ frameName($0) })
+    }
+}
 
 /// An enum used to identify the different types of frame parsed by the ID3TagEditor.
 /// This must be used to acces the frame data as identifier inside the dictionary
 /// of frame in the `ID3tag` `frames` properties.
 public enum FrameName: Equatable, Hashable, CaseIterable {
     public static var allCases: [FrameName] {
+        let frameNamesWithLocalizedContent = FrameNamesWithLocalizedContent().get()
         return [
             .title,
             .album,
@@ -41,6 +46,7 @@ public enum FrameName: Equatable, Hashable, CaseIterable {
             .genre,
             .discPosition,
             .trackPosition,
+            .initialKey,
             .recordingDayMonth,
             .recordingYear,
             .recordingHourMinute,
@@ -102,6 +108,8 @@ public enum FrameName: Equatable, Hashable, CaseIterable {
     case discPosition
     /// Track Position frame name.
     case trackPosition
+    /// Initial key frame name.
+    case initialKey
     /// RecordingDayMonth frame name. Valid only for tag version 2.3 and 2.2. For version 2.4 use RecordingDateTime.
     case recordingDayMonth
     /// RecordingYear frame name. Valid only for tag version 2.3 and 2.2. For version 2.4 use RecordingDateTime.
